// packages/clerk/src/operations/platform/PlatformBanUser.ts
export const PlatformBanUserInput = /*@__PURE__*/ /*#__PURE__*/ Schema.Struct({
  applicationID: Schema.String.pipe(T.PathParam()),
  envOrInsID: Schema.String.pipe(T.PathParam()),
  userID: Schema.String.pipe(T.PathParam()),
}).pipe(
  T.Http({
    method: "POST",
    path: "/platform/applications/{applicationID}/instances/{envOrInsID}/users/{userID}/ban",
  }),
);
export type PlatformBanUserInput = typeof PlatformBanUserInput.Type;
// ...
export type PlatformBanUserOutput = typeof PlatformBanUserOutput.Type;
// ...
export const PlatformBanUser = /*@__PURE__*/ /*#__PURE__*/ API.make(() => ({
  inputSchema: PlatformBanUserInput,
  outputSchema: PlatformBanUserOutput,
  errors: [BadRequest, Forbidden, NotFound] as const,
}));