// packages/clerk/src/operations/platform/PlatformListInstanceUsers.ts
export const PlatformListInstanceUsersInput =
  /*@__PURE__*/ /*#__PURE__*/ Schema.Struct({
    applicationID: Schema.String.pipe(T.PathParam()),
    envOrInsID: Schema.String.pipe(T.PathParam()),
    query: Schema.optional(Schema.String),
    order_by: Schema.optional(Schema.String),
    limit: Schema.optional(Schema.Number),
    offset: Schema.optional(Schema.Number),
  }).pipe(
    T.Http({
      method: "GET",
      path: "/platform/applications/{applicationID}/instances/{envOrInsID}/users",
    }),
  );
export type PlatformListInstanceUsersInput =
  typeof PlatformListInstanceUsersInput.Type;
// ...
export type PlatformListInstanceUsersOutput =
  typeof PlatformListInstanceUsersOutput.Type;
// ...
export const PlatformListInstanceUsers = /*@__PURE__*/ /*#__PURE__*/ API.make(
  () => ({
    inputSchema: PlatformListInstanceUsersInput,
    outputSchema: PlatformListInstanceUsersOutput,
    errors: [BadRequest, Forbidden, NotFound, UnprocessableEntity] as const,
  }),
);